feat: template attributes can now use ParsedArgumentValue instead of just the raw type
